# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion Pénitentiaire API",
        "description": "API REST pour le recensement des détenus, le suivi des condamnations et les statistiques pénitentiaires.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/detenus": {
            "get": {
                "tags": [
                    "Détenus"
                ],
                "summary": "Lister / rechercher les détenus",
                "description": "Recherche multicritère par nom, numéro Nina, date d'entrée ou statut.",
                "operationId": "0afa7ee2a62a43720909cb1cc54505ae",
                "parameters": [
                    {
                        "name": "nom",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Diallo"
                    },
                    {
                        "name": "nina",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "NINA-2026001"
                    },
                    {
                        "name": "date_entree",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        },
                        "example": "2024-01-10"
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "present",
                                "libere",
                                "decede"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste paginée des détenus"
                    }
                }
            },
            "post": {
                "tags": [
                    "Détenus"
                ],
                "summary": "Enregistrer un détenu",
                "operationId": "2d275f5e848e761c9066ab2a9762f820",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DetenuInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Détenu créé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/detenus/{detenu}": {
            "get": {
                "tags": [
                    "Détenus"
                ],
                "summary": "Fiche de synthèse d'un détenu",
                "description": "Affiche les informations personnelles, condamnations, temps écoulé et temps restant.",
                "operationId": "0e2a37f6725e673532be7a5fc9a1b450",
                "parameters": [
                    {
                        "name": "detenu",
                        "in": "path",
                        "description": "ID ou NINA selon votre Route Model Binding",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "NINA-2026001"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fiche détenu"
                    },
                    "404": {
                        "description": "Détenu introuvable"
                    }
                }
            },
            "put": {
                "tags": [
                    "Détenus"
                ],
                "summary": "Modifier un détenu",
                "operationId": "bf44a52523d3784842a2277494238726",
                "parameters": [
                    {
                        "name": "detenu",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "NINA-2026001"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DetenuInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Détenu mis à jour"
                    },
                    "404": {
                        "description": "Détenu introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Détenus"
                ],
                "summary": "Supprimer un détenu",
                "operationId": "911986e1da935d8bc8b88b9ca292a346",
                "parameters": [
                    {
                        "name": "detenu",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "NINA-2026001"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détenu supprimé"
                    },
                    "404": {
                        "description": "Détenu introuvable"
                    }
                }
            }
        },
...
        "/detenus/listes/sorties-prevues": {
            "get": {
                "tags": [
                    "Listes"
                ],
                "summary": "Sorties prévues pour un mois donné",
                "operationId": "21330537137bc8c47c3bbbe4bf0b70f3",
                "parameters": [
                    {
                        "name": "mois",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    },
                    {
                        "name": "annee",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 2026
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détenus dont la fin de peine est prévue ce mois-ci"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "DetenuInput": {
                "required": [
                    "matricule_ou_nina",
                    "nom",
                    "prenom",
                    "date_naissance",
                    "sexe"
                ],
                "properties": {
                    "matricule_ou_nina": {
                        "type": "string",
                        "example": "NINA-2026001"
                    },
                    "nom": {
                        "type": "string",
                        "example": "Diallo"
                    },
                    "prenom": {
                        "type": "string",
                        "example": "Amadou"
                    },
                    "date_naissance": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-05-15"
                    },
                    "sexe": {
                        "type": "string",
                        "example": "M",
                        "enum": [
                            "M",
                            "F"
                        ]
                    },
                    "photo": {
                        "type": "string",
                        "example": "photos/detenu.jpg",
                        "nullable": true
                    },
                    "statut": {
                        "type": "string",
                        "example": "present",
                        "enum": [
                            "present",
                            "libere",
                            "decede"
                        ]
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}